created new class for gas station API
