+++ src/main/java/com/detroitlabs/FinalProject/controller/TripController.java
@@ -7,16 +7,13 @@ import com.detroitlabs.FinalProject.service.TripService;
 import com.detroitlabs.FinalProject.service.YelpService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.util.ArrayList;
-import java.util.List;
 
 @Controller
 public class TripController {
@@ -31,7 +28,7 @@ public class TripController {
     private YelpService yelpService;
 
     @Autowired
-    private Stations stations;
+    private StationsContents stationsContents;
 
     @Autowired
     private TripService tripService;
@@ -44,6 +41,9 @@ public class TripController {
 
     @Autowired
     StationsWrapper stationsWrapper;
+
+    @Autowired
+   Stations stations;
 
     @Value("${GOOGLE_MAPS_KEY}")
     private String googleMapsKey;
@@ -96,6 +96,10 @@ public class TripController {
        modelMap.put("filteredCityNames",filteredCityNames);
        modelMap.put("tripCityPlaces", tripCityPlaces.getTripCityPlaces());
        modelMap.put("googleMapsKey", googleMapsKey);
+
+       //Gas Station Info
+
+
 
         return "showtrip";
     }
