+++ src/main/java/com/detroitlabs/FinalProject/service/TripService.java
@@ -1,6 +1,5 @@
 package com.detroitlabs.FinalProject.service;
 
-import com.detroitlabs.FinalProject.model.Stations;
 import com.detroitlabs.FinalProject.model.StationsWrapper;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
